+++ src/app/(page)/(auth)/login/page.tsx
@@ -1,5 +1,5 @@
 import { Metadata } from "next";
-import LoginForm from "../../../../components/auth/LoginForm";
+import LoginForm from "../../../../components/(auth)/LoginForm";
 
 export const metadata: Metadata = {
   title: "ログイン | ScoreGraph",
