+++ src/components/(profile)/ProfileUpdate.tsx
@@ -6,7 +6,7 @@ import { CircleUserRound } from "lucide-react";
 import Image from "next/image";
 import { ProfileData, ProfileUpdateData } from "@/type/userType";
 import { ROUTES } from "@/constants";
-import BackButton from "../general/BackButton";
+import BackMypageLink from "../general/BackMypageLink";
 
 interface ProfileEditFormProps {
   initialData: ProfileData;
@@ -144,7 +144,7 @@ export default function ProfileUpdate({ initialData }: ProfileEditFormProps) {
       </div>
 
       <div className="flex justify-between p-2">
-        <BackButton />
+        <BackMypageLink />
         <button
           type="submit"
           disabled={isSubmitting}
