+++ src/components/(target)/TargetUpsert.tsx
@@ -3,10 +3,9 @@
 import React, { useState } from "react";
 import { Subject } from "@/type/testType";
 import { TargetData } from "@/type/targetType";
-import { useRouter } from "next/navigation";
-import { toast } from "react-hot-toast"; // トースト通知用（必要に応じてインストール）
-import { ROUTES } from "@/constants";
+import { toast } from "react-hot-toast";
 import { displaySubjectName } from "@/lib/display";
+import BackMypageLink from "../general/BackMypageLink";
 
 type Props = {
   subject: Subject;
@@ -14,7 +13,6 @@ type Props = {
   existingData: Record<string, TargetData>;
 };
 
-// 月の表示用フォーマット
 const formatMonth = (month: string) => {
   const [year, monthNum] = month.split("-");
   return `${year}年${monthNum}月`;
@@ -25,18 +23,15 @@ export default function TargetUpsert({
   targetMonthTemplate,
   existingData,
 }: Props) {
-  const router = useRouter();
   const [isSubmitting, setIsSubmitting] = useState(false);
-  const [activeMonth, setActiveMonth] = useState<string | null>(null);
+  const [editingMonth, setEditingMonth] = useState<string | null>(null);
 
-  // フォーム入力値の状態
   const [formValues, setFormValues] = useState<{
     [month: string]: {
       targetPercentage: number;
       targetMemo: string;
     };
   }>(() => {
-    // 既存データからフォーム初期値を設定
     const initialValues: {
       [month: string]: { targetPercentage: number; targetMemo: string };
     } = {};
@@ -49,7 +44,7 @@ export default function TargetUpsert({
         };
       } else {
         initialValues[month] = {
-          targetPercentage: 70, // デフォルト値
+          targetPercentage: 70,
           targetMemo: "",
         };
       }
@@ -58,7 +53,6 @@ export default function TargetUpsert({
     return initialValues;
   });
 
-  // 入力値を更新する関数
   const handleInputChange = (
     month: string,
     field: "targetPercentage" | "targetMemo",
@@ -73,7 +67,32 @@ export default function TargetUpsert({
     }));
   };
 
-  // データを保存する関数
+  const handleEdit = (month: string) => {
+    setEditingMonth(month);
+  };
+
+  const handleCancel = (month: string) => {
+    // 編集前の値に戻す
+    if (existingData[month]) {
+      setFormValues((prev) => ({
+        ...prev,
+        [month]: {
+          targetPercentage: existingData[month].targetPercentage,
+          targetMemo: existingData[month].targetMemo,
+        },
+      }));
+    } else {
+      setFormValues((prev) => ({
+        ...prev,
+        [month]: {
+          targetPercentage: 70,
+          targetMemo: "",
+        },
+      }));
+    }
+    setEditingMonth(null);
+  };
+
   const handleSubmit = async (month: string) => {
     setIsSubmitting(true);
 
@@ -95,11 +114,10 @@ export default function TargetUpsert({
 
       if (result.success) {
         toast.success(`${formatMonth(month)}の目標を保存しました`);
-        // 既存データを更新
         existingData[month] = {
           ...existingData[month],
           id: result.targetId,
-          userId: 1, // ダミー値
+          userId: 1,
           subject,
           targetPercentage: formValues[month].targetPercentage,
           targetMonth: month,
@@ -107,9 +125,7 @@ export default function TargetUpsert({
           createdAt: new Date(),
           updatedAt: new Date(),
         };
-
-        // アクティブな月をリセット
-        setActiveMonth(null);
+        setEditingMonth(null);
       } else {
         toast.error(
           `保存エラー: ${result.error || "不明なエラーが発生しました"}`
@@ -123,148 +139,106 @@ export default function TargetUpsert({
     }
   };
 
-  // カードページに戻る
-  const handleBack = () => {
-    router.push(`${ROUTES.MYPAGE}?tab=cards`);
-  };
-
   return (
-    <div className="bg-white rounded-lg shadow p-6 mb-16">
-      <h2 className="text-lg font-semibold mb-4">
-        {displaySubjectName(subject)}の目標設定
+    <div className="bg-white rounded-lg shadow p-6 mb-16 text-gray-700">
+      <h2 className="text-lg font-bold mb-4">
+        折れ線グラフ（目標）
+        <span className="ml-4">{displaySubjectName(subject)}</span>
       </h2>
 
       <div className="mb-6">
-        <p className="text-sm text-gray-600 mb-2">
-          各月の目標を設定してください。目標は後からいつでも変更できます。
+        <p className="text-sm text-gray-400 mb-2">
+          折れ線グラフを作成します。目標にする「%」を設定してください。目標は後からいつでも変更できます。
         </p>
       </div>
 
-      <div className="space-y-6">
-        {targetMonthTemplate.map((month) => (
-          <div key={month} className="border rounded-lg p-4">
-            <div className="flex justify-between items-center mb-3">
-              <h3 className="font-medium">{formatMonth(month)}</h3>
+      {targetMonthTemplate.map((month) => (
+        <div key={month}>
+          <div className="flex justify-between items-center mb-3">
+            <h3 className="font-bold">{formatMonth(month)}</h3>
 
-              {existingData[month] ? (
-                <span className="text-xs bg-green-100 text-green-800 px-2 py-1 rounded">
-                  設定済み
-                </span>
+            <button
+              onClick={() => handleEdit(month)}
+              className={
+                "text-xs px-2 py-1 rounded-md bg-gray-100 text-gray-500"
+              }
+              disabled={editingMonth === month}
+            >
+              {editingMonth === month ? "編集中" : "設定する"}
+            </button>
+          </div>
+
+          <div className="space-y-2">
+            <div>
+              <label className="block mb-1">目標（%）</label>
+              {editingMonth === month ? (
+                <input
+                  type="text"
+                  inputMode="numeric"
+                  pattern="[0-9]*"
+                  min="0"
+                  max="100"
+                  value={formValues[month].targetPercentage}
+                  onChange={(e) =>
+                    handleInputChange(month, "targetPercentage", e.target.value)
+                  }
+                  className="w-full px-3 py-2 rounded-lg bg-gray-100 focus:outline-none"
+                />
               ) : (
-                <span className="text-xs bg-gray-100 text-gray-600 px-2 py-1 rounded">
-                  未設定
-                </span>
+                <p className="px-3 py-2 font-bold text-blue-500">
+                  {formValues[month].targetPercentage}%
+                </p>
               )}
             </div>
 
-            {activeMonth === month ? (
-              <div className="space-y-4">
-                <div>
-                  <label className="block text-sm font-medium text-gray-700 mb-1">
-                    目標正答率 (%)
-                  </label>
-                  <input
-                    type="number"
-                    min="0"
-                    max="100"
-                    value={formValues[month].targetPercentage}
-                    onChange={(e) =>
-                      handleInputChange(
-                        month,
-                        "targetPercentage",
-                        e.target.value
-                      )
-                    }
-                    className="w-full px-3 py-2 border rounded-md text-sm"
-                  />
-                </div>
+            <div>
+              <label className="block text-sm mb-1 text-gray-500">メモ</label>
+              {editingMonth === month ? (
+                <textarea
+                  value={formValues[month].targetMemo}
+                  onChange={(e) =>
+                    handleInputChange(month, "targetMemo", e.target.value)
+                  }
+                  className="w-full px-3 py-2 rounded-lg bg-gray-100 focus:outline-none"
+                  rows={3}
+                  placeholder="例）時間制限は設けず、現状どれくらいの精度で解けるのか確認することがテーマです。..."
+                />
+              ) : (
+                <p className="px-3 pt-2">
+                  {formValues[month].targetMemo || (
+                    <span className="text-gray-400 text-sm">未記入</span>
+                  )}
+                </p>
+              )}
+            </div>
 
-                <div>
-                  <label className="block text-sm font-medium text-gray-700 mb-1">
-                    メモ
-                  </label>
-                  <textarea
-                    value={formValues[month].targetMemo}
-                    onChange={(e) =>
-                      handleInputChange(month, "targetMemo", e.target.value)
-                    }
-                    className="w-full px-3 py-2 border rounded-md text-sm"
-                    rows={3}
-                    placeholder="目標達成のためのメモや戦略など"
-                  ></textarea>
-                </div>
-
-                <div className="flex justify-end space-x-2">
-                  <button
-                    type="button"
-                    onClick={() => setActiveMonth(null)}
-                    className="px-3 py-1 bg-gray-100 text-gray-700 rounded-md text-sm"
-                    disabled={isSubmitting}
-                  >
-                    キャンセル
-                  </button>
-
-                  <button
-                    type="button"
-                    onClick={() => handleSubmit(month)}
-                    className="px-3 py-1 bg-blue-600 text-white rounded-md text-sm disabled:opacity-50"
-                    disabled={isSubmitting}
-                  >
-                    {isSubmitting ? "保存中..." : "保存"}
-                  </button>
-                </div>
-              </div>
-            ) : (
-              <div>
-                {existingData[month] ? (
-                  <div className="space-y-2">
-                    <div className="flex items-center">
-                      <span className="text-sm font-medium text-gray-500 mr-2">
-                        目標正答率:
-                      </span>
-                      <span className="text-sm">
-                        {formValues[month].targetPercentage}%
-                      </span>
-                    </div>
-
-                    {formValues[month].targetMemo && (
-                      <div>
-                        <span className="text-sm font-medium text-gray-500">
-                          メモ:
-                        </span>
-                        <p className="text-sm text-gray-700 mt-1">
-                          {formValues[month].targetMemo}
-                        </p>
-                      </div>
-                    )}
-                  </div>
-                ) : (
-                  <p className="text-sm text-gray-500">
-                    この月の目標はまだ設定されていません。
-                  </p>
-                )}
-
+            {editingMonth === month && (
+              <div className="flex justify-end gap-4 mr-2">
                 <button
                   type="button"
-                  onClick={() => setActiveMonth(month)}
-                  className="mt-3 px-3 py-1 bg-gray-100 text-gray-700 rounded-md text-sm"
+                  onClick={() => handleCancel(month)}
+                  className="text-gray-500 text-sm"
+                  disabled={isSubmitting}
                 >
-                  {existingData[month] ? "編集" : "設定する"}
+                  キャンセル
+                </button>
+                <button
+                  type="button"
+                  onClick={() => handleSubmit(month)}
+                  className="text-blue-500 font-bold text-sm"
+                  disabled={isSubmitting}
+                >
+                  {isSubmitting ? "保存中..." : "完了"}
                 </button>
               </div>
             )}
           </div>
-        ))}
-      </div>
+          <div className="border-b border-gray-300 my-6"></div>
+        </div>
+      ))}
 
-      <div className="mt-8 flex justify-between">
-        <button
-          type="button"
-          onClick={handleBack}
-          className="px-4 py-2 bg-gray-100 text-gray-700 rounded-md"
-        >
-          カード一覧に戻る
-        </button>
+      <div className="p-2 mt-6">
+        <BackMypageLink />
       </div>
     </div>
   );
